requirements.txt
black==22.3.0
py-solc-x==1.1.1
web3==5.29.0
